Add success toast for role/user update. Refactor cross page toast.
